test(home): consolidate scenarios
