+++ e2e/home.spec.ts
@@ -1,35 +1,9 @@
 import { expect, test, type Page } from '@playwright/test';
-import { resetAppData, selectDays, selectWorkouts } from './helpers';
-
-const DAYS = ['SUNDAY', 'MONDAY', 'TUESDAY', 'WEDNESDAY', 'THURSDAY', 'FRIDAY', 'SATURDAY'];
-
-async function openHomePage(page: Page) {
-  await resetAppData(page);
-  await page.goto('/');
-  await expect(page.getByText("Today's Workouts")).toBeVisible();
-}
-
-async function openPlansPage(page: Page) {
-  await page.getByRole('button', { name: 'Plans' }).click();
-  await expect(page.getByRole('heading', { name: 'Workout Plans' })).toBeVisible();
-}
-
-async function createPlan(page: Page, planName: string, workoutName: string) {
-  await openPlansPage(page);
-  await page.getByRole('button', { name: 'add workout' }).click();
-  await expect(page.getByRole('heading', { name: 'Create Workout Plan' })).toBeVisible();
-
-  await page.getByLabel('Plan Name').fill(planName);
-  await selectWorkouts(page, [workoutName]);
-  await selectDays(page, DAYS);
-  await page.getByRole('button', { name: 'Save' }).click();
-
-  await expect(page.getByRole('heading', { name: 'Create Workout Plan' })).toBeHidden();
-  await expect(page.getByText(planName)).toBeVisible();
-}
+import { createPlan, fillWorkoutSet, openHomePage, openWorkout } from './home-scenario-helpers';
 
 async function clearAllPlans(page: Page) {
-  await openPlansPage(page);
+  await page.getByRole('button', { name: 'Plans' }).click();
+  await expect(page.getByRole('heading', { name: 'Workout Plans' })).toBeVisible();
 
   const initialCount = await page.getByRole('listitem').count();
   for (let i = 0; i < initialCount; i++) {
@@ -59,5 +33,101 @@ test.describe('Home', () => {
 
     await expect(page.getByText("Today's Workouts")).toBeVisible();
     await expect(page.getByRole('listitem')).toHaveCount(0);
+  });
+
+  test('3.4 open the workout tracker for a workout', async ({ page }) => {
+    const planName = `Scenario Plan ${Date.now()}`;
+
+    await openHomePage(page);
+    await createPlan(page, planName, 'Running');
+    await page.getByRole('button', { name: 'Home' }).click();
+
+    await openWorkout(page, 'Running');
+    await expect(page.getByRole('heading', { name: 'Workout Details' })).toBeVisible();
+  });
+
+  test('3.5 log a workout session and save', async ({ page }) => {
+    const planName = `Scenario Plan ${Date.now()}`;
+
+    await openHomePage(page);
+    await createPlan(page, planName, 'Running');
+    await page.getByRole('button', { name: 'Home' }).click();
+
+    await openWorkout(page, 'Running');
+    await fillWorkoutSet(page, 1, [30]);
+    await expect(page.getByRole('button', { name: /^save$/i })).toBeEnabled();
+    await page.getByRole('button', { name: /^save$/i }).click();
+
+    await expect(page.getByRole('heading', { name: 'Workout Details' })).toBeHidden();
+    await expect(page.getByTestId('CheckCircleOutlineIcon')).toBeVisible();
+    await expect(page.getByText('Running', { exact: true })).toHaveCSS('color', 'rgba(255, 255, 255, 0.5)');
+  });
+
+  test('3.6 save is disabled until all required fields are filled', async ({ page }) => {
+    const planName = `Scenario Plan ${Date.now()}`;
+
+    await openHomePage(page);
+    await createPlan(page, planName, 'Running');
+    await page.getByRole('button', { name: 'Home' }).click();
+
+    await openWorkout(page, 'Running');
+    await expect(page.getByRole('button', { name: /^save$/i })).toBeDisabled();
+
+    await page.getByRole('dialog').locator('input').nth(0).fill('30');
+    await page.waitForTimeout(400);
+    await expect(page.getByRole('button', { name: /^save$/i })).toBeEnabled();
+  });
+
+  test('3.7 add multiple sets before saving', async ({ page }) => {
+    const planName = `Scenario Plan ${Date.now()}`;
+
+    await openHomePage(page);
+    await createPlan(page, planName, 'Running');
+    await page.getByRole('button', { name: 'Home' }).click();
+
+    await openWorkout(page, 'Running');
+    await fillWorkoutSet(page, 1, [30]);
+
+    const addSetButton = page.getByRole('button', { name: 'Add Set' });
+    const removeSetButton = page.getByRole('button', { name: 'Remove Set' });
+
+    await expect(addSetButton).toBeEnabled();
+    await addSetButton.click();
+
+    await expect(page.getByText('Set 2')).toBeVisible();
+    await expect(addSetButton).toBeDisabled();
+    await expect(removeSetButton).toBeEnabled();
+
+    await fillWorkoutSet(page, 2, [32]);
+    await expect(addSetButton).toBeEnabled();
+
+    await removeSetButton.click();
+    await expect(page.getByText('Set 2')).toHaveCount(0);
+  });
+
+  test('3.8 remove set is disabled when only one set remains', async ({ page }) => {
+    const planName = `Scenario Plan ${Date.now()}`;
+
+    await openHomePage(page);
+    await createPlan(page, planName, 'Running');
+    await page.getByRole('button', { name: 'Home' }).click();
+
+    await openWorkout(page, 'Running');
+    await expect(page.getByRole('button', { name: 'Remove Set' })).toBeDisabled();
+  });
+
+  test('3.9 previously recorded data is shown in the tracker', async ({ page }) => {
+    const planName = `History Plan ${Date.now()}`;
+
+    await openHomePage(page);
+    await createPlan(page, planName, 'Push-ups');
+    await page.getByRole('button', { name: 'Home' }).click();
+
+    await page.getByText('Push-ups', { exact: true }).click();
+    await expect(page.getByRole('heading', { name: 'Workout Details' })).toBeVisible();
+    await expect(page.getByText('Previously Recorded')).toBeVisible();
+    await expect(page.getByText('Set 1:')).toBeVisible();
+    await expect(page.getByText('10 Reps for 30Sec')).toBeVisible();
+    await expect(page.locator('.recharts-responsive-container svg.recharts-surface').first()).toBeVisible();
   });
 });
